добавил структуры

+++ Artur_Skomarohov_HW2_FigureCalc/Triangle.cs
@@ -6,7 +6,7 @@ namespace FigureCalc
 {
     public class Triangle
     {
-        public double SideNumber { get; set; }
+        public int SideNumber { get; set; }
         public double SideA { get; set; }
         public double SideB { get; set; }
         public double SideC { get; set; }
